move added functions to separate file
